Improve login page design
The login page had an inconsistent look and feel compared to the rest of the application, which could confuse users. To address this, we integrated the login page with the main application template, added a centred login form frame, unified the form element styles, and vertically centered the login content. These changes aimed to create a more visually cohesive and user-friendly login experience that feels like a natural part of the overall application.